refactor - move code from CreateDirectMessage to MembersModal
